fix problem with profile account
