+++ src/widget/Sidebar/model/selectors/getSidebarItems.ts
@@ -1,0 +1,46 @@
+import { createSelector } from '@reduxjs/toolkit';
+import { getUserAuthData } from 'enteties/User';
+import { RoutePath } from 'shared/config/routeConfig/routeConfig';
+import MainIcon from 'shared/assets/icons/main-20-20.svg';
+import AboutIcon from 'shared/assets/icons/about-20-20.svg';
+import ProfileIcon from 'shared/assets/icons/profile-20-20.svg';
+import ArticleIcon from 'shared/assets/icons/article-20-20.svg';
+import { SidebarItemType } from '../types/sidebar';
+
+export const getSidebarItems = createSelector(
+  // Получаем данные о пользователе
+  getUserAuthData,
+  (userData) => {
+    const sidebarItemList: SidebarItemType[] = [
+      {
+        path: RoutePath.main,
+        Icon: MainIcon,
+        text: 'Главная',
+      },
+      {
+        path: RoutePath.about,
+        Icon: AboutIcon,
+        text: 'О сайте',
+      },
+    ];
+    // Если пользователь авторизован, то добавляем следующие ссылки
+    if (userData) {
+      sidebarItemList.push(
+        {
+          path: RoutePath.profile + userData.id,
+          Icon: ProfileIcon,
+          text: 'Профиль',
+          authOnly: true,
+        },
+        {
+          path: RoutePath.articles,
+          Icon: ArticleIcon,
+          text: 'Статьи',
+          authOnly: true,
+        },
+      );
+    }
+
+    return sidebarItemList;
+  },
+);
